+++ starkinfra/indirectparticipant/__init__.py
@@ -1,1 +1,1 @@
-from .__indirectparticipant import get, create
+from .__indirectparticipant import create, get, delete