+++ starkinfra/issuingholder/__issuingholder.py
@@ -20,8 +20,7 @@ class IssuingHolder(Resource):
     - created [datetime.datetime]: creation datetime for the IssuingHolder. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     """
 
-    def __init__(self, id=None, external_id=None, name=None, rules=None, status=None, tags=None, tax_id=None,
-                 updated=None, created=None):
+    def __init__(self, name, tax_id, external_id, id=None,  rules=None, status=None, tags=None, updated=None, created=None):
         Resource.__init__(self, id=id)
         self.name = name
         self.tax_id = tax_id