+++ starkinfra/issuinginvoice/__init__.py
@@ -1,3 +1,3 @@
-from . import log
-from .log.__log import Log
 from .__issuinginvoice import create, get, query, page
+from .log.__log import Log
+from . import log