+++ tests/utils/__init__.py
@@ -1,2 +1,0 @@
-from . import endToEndId
-from .endToEndId import get_end_to_end_id